+++ main_test.go
@@ -64,29 +64,29 @@ func TestReadTokens(t *testing.T) {
 func TestEnv(t *testing.T) {
 	t.Run("plus", func(t *testing.T) {
 		env := NewEnv()
-		env.update("+", BinaryFunc{func(x, y int) int { return x + y }})
+		env.update("+", BinaryFunc{func(x, y float64) float64 { return x + y }})
 
 		fn := mustFindBinaryFunc(t, env, "+")
 
 		got := fn.f(1, 3)
-		want := 4
+		want := 4.0
 
 		if got != want {
-			t.Errorf("got %d want %d", got, want)
+			t.Errorf("got %f want %f", got, want)
 		}
 	})
 
 	t.Run("minus", func(t *testing.T) {
 		env := NewEnv()
-		env.update("-", BinaryFunc{func(x, y int) int { return x - y }})
+		env.update("-", BinaryFunc{func(x, y float64) float64 { return x - y }})
 
 		fn := mustFindBinaryFunc(t, env, "-")
 
 		got := fn.f(3, 1)
-		want := 2
+		want := 2.0
 
 		if got != want {
-			t.Errorf("got %d want %d", got, want)
+			t.Errorf("got %f want %f", got, want)
 		}
 	})
